Update pxt.json, main.blocks, main.ts, tilemap.g.jres, tilemap.g.ts, jumpable.ts
